% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titegboin.R
\name{get.oc.tite.gBOIN}
\alias{get.oc.tite.gBOIN}
\title{Obtain the operating characteristics of the Time-to-event gBOIN design for late-onset toxicity grades by simulating trials.}
\usage{
get.oc.tite.gBOIN(
  target,
  prob,
  ncohort,
  cohortsize,
  maxt = 1,
  prior.p = NA,
  accrual = 3,
  maxpen = 0.5,
  alpha = 0.5,
  startdose = 1,
  cutoff.eli = 0.95,
  ntrial = 1000,
  seed = seed
)
}
\arguments{
\item{target}{the target ETS}

\item{prob}{a matrix containing the true toxicity probabilities of each grade x the investigational dose levels.}

\item{ncohort}{the total number of cohorts}

\item{cohortsize}{the cohort size}

\item{maxt}{the maximum follow-up time}

\item{prior.p}{a vector of length 3, which specifies the prior probability that the time to toxicity lies
inside the time interval (0,\code{maxt}/3), (\code{maxt}/3,\code{2maxt}/3), (\code{2maxt}/3,1).
The default value is \code{prior.p=c(1/3,1/3,1/3)}.}

\item{accrual}{the accrual rate, i.e., the number of patients accrued in 1 unit of time,}

\item{maxpen}{the maximum proportion of pending patients}

\item{alpha}{a number from (0,1) that controls alpha*100\% events in (0, 1/2T).
The default is \code{alpha=0.5}.}

\item{startdose}{the starting dose level for the trial}

\item{cutoff.eli}{the cutoff to eliminate an overly toxic dose for safety.
We recommend the default value of (\code{cutoff.eli=0.95}) for general use.}

\item{ntrial}{the total number of trials to be simulated.}

\item{seed}{set seed}
}
\value{
\code{get.oc.tite.gBOIN()} returns the operating characteristics of the TITE-gBOIN design as a data frame,
including: (1) selection percentage at each dose level (\code{selpercent}),
(2) the number of patients treated at each dose level (\code{nptsdose}),
(3) the number of toxicities observed at each dose level (\code{ntoxdose}),
(4) the average number of toxicities (\code{totaltox}),
(5) the average number of patients (\code{totaln}),
(6) the percentage of early stopping without selecting the MTD (\code{pctearlystop}).
(7) the average trial duration needed for the trial based on the TITE-gBOIN design (\code{duration})
(8)the standard deviation of trial duration needed for the trial based on the TITE-gBOIN design (\code{sdduration})
}
\description{
Obtain the operating characteristics of the Time-to-event gBOIN design for late-onset toxicity grades by simulating trials.
}
\details{
This function generates he operating characteristics of the Time-To-Event Bayesian Optimal
Interval design for toxicity grades(TITE-gBOIN)
with delayed toxicities grades by simulating trials under the prespecified true ETS of the investigational doses.
}
