% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titegboin-package.R
\docType{package}
\name{titegboin-package}
\alias{titegboin}
\alias{titegboin-package}
\title{titegboin: Obtain the operating characteristics of the Time-to-event gBOIN design for late-onset toxicity grades by simulating trials.}
\description{
Obtain the operating characteristics of the Time-to-event gBOIN design for late-onset toxicity grades by simulating trials.
}
\keyword{internal}
